[Refactor] 뷰 분리 (#14)
